Moved `functions` package init logic to separate file
