+++ packages/functions/src/lib/env.ts
@@ -1,0 +1,8 @@
+import {projectID} from 'firebase-functions/params';
+
+// TODO: Make region an environment variable.
+const FIREBASE_FUNCTIONS_REGION = 'us-central1';
+
+export function getFunctionsBaseUrl(): string {
+  return `https://${FIREBASE_FUNCTIONS_REGION}-${projectID.value()}.cloudfunctions.net`;
+}
